clinic_reception/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
docx2pdf==0.1.8
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==6.0.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
pywin32==310
pyzmq==26.4.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
temporalio==1.3.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
wcwidth==0.2.13